final commit

+++ holdings-tab.go
@@ -13,33 +13,41 @@ import (
 )
 
 func (app *Config) holdingsTab() *fyne.Container {
+	app.Holdings = app.getHoldingSlice()
 	app.HoldingsTable = app.getHoldingsTable()
 
-	holdingsContainer := container.NewVBox(app.HoldingsTable)
+	holdingsContainer := container.NewBorder(
+		nil,
+		nil,
+		nil,
+		nil,
+		container.NewAdaptiveGrid(1, app.HoldingsTable),
+	)
 
 	return holdingsContainer
 }
 
 func (app *Config) getHoldingsTable() *widget.Table {
-	data := app.getHoldingSlice()
-	app.Holdings = data
 	t := widget.NewTable(
 		func() (rows int, cols int) {
-			return len(data), len(data[0])
+			return len(app.Holdings), len(app.Holdings[0])
 		},
 		func() fyne.CanvasObject {
 			ctr := container.NewVBox(widget.NewLabel(""))
 			return ctr
 		},
 		func(i widget.TableCellID, o fyne.CanvasObject) {
-			if i.Col == (len(data[0])-1) && i.Row != 0 {
+			if i.Col == (len(app.Holdings[0])-1) && i.Row != 0 {
 				// last cell - put in a button
 				w := widget.NewButtonWithIcon("Delete", theme.DeleteIcon(), func() {
 					dialog.ShowConfirm("Delete?", "", func(deleted bool) {
-						id, _ := strconv.Atoi(data[i.Row][0].(string)) // Cast to string
-						err := app.DB.DeleteHolding(int64(id))
-						if err != nil {
-							app.ErrorLog.Println(err)
+						if deleted {
+
+							id, _ := strconv.Atoi(app.Holdings[i.Row][0].(string)) // Cast to string
+							err := app.DB.DeleteHolding(int64(id))
+							if err != nil {
+								app.ErrorLog.Println(err)
+							}
 						}
 
 						// Refresh holding table
@@ -53,7 +61,7 @@ func (app *Config) getHoldingsTable() *widget.Table {
 			} else {
 				// we're just putting in textual information
 				o.(*fyne.Container).Objects = []fyne.CanvasObject{
-					widget.NewLabel(data[i.Row][i.Col].(string)),
+					widget.NewLabel(app.Holdings[i.Row][i.Col].(string)),
 				} // Casting
 			}
 		},
@@ -61,7 +69,7 @@ func (app *Config) getHoldingsTable() *widget.Table {
 
 	colWidths := []float32{50, 200, 200, 200, 100}
 
-	for i := 0; i < len(data[0]); i++ {
+	for i := 0; i < len(app.Holdings[0]); i++ {
 		t.SetColumnWidth(i, colWidths[i])
 	}
 
@@ -83,7 +91,9 @@ func (app *Config) getHoldingSlice() [][]interface{} {
 
 		currentRow = append(currentRow, strconv.FormatInt(x.ID, 10))
 		currentRow = append(currentRow, fmt.Sprintf("%d toz", x.Amount))
-		currentRow = append(currentRow, fmt.Sprintf("%2f", float32(x.PurchasePrice)/100))
+		currentAmount := float32(x.PurchasePrice)
+		currentAmount = currentAmount / 100
+		currentRow = append(currentRow, fmt.Sprintf("$%.2f", currentAmount))
 		currentRow = append(currentRow, x.PurchaseDate.Format("2006-01-02"))
 		currentRow = append(currentRow, widget.NewButton("Delete", func() {}))
